# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZQAI API",
    "description": "REST API for the ZQAI portal (news, events, managers, departments, compliance officer, reception schedule, file uploads and user/auth management). Backend: Express 5 + PostgreSQL.\n\nNOTE: BearerAuth is documented as the intended security scheme for state-changing and administrative operations, but the current implementation does NOT enforce any authentication middleware. The live scan is expected to reveal this as a broken-authentication / broken-function-level-authorization finding.",
    "version": "0.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development server" }
  ],
...
  "paths": {
...
    "/api/managers": {
      "get": {
        "operationId": "getManagers",
        "summary": "List managers",
        "tags": ["Managers"],
        "parameters": [
          { "name": "lang", "in": "query", "required": false, "schema": { "type": "string", "enum": ["ru", "kk", "en"] } }
        ],
        "responses": {
          "200": {
            "description": "List of managers",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Manager" } } } }
          },
          "500": { "description": "Failed to fetch managers" }
        }
      },
      "post": {
        "operationId": "createManager",
        "summary": "Create a manager (should be authenticated)",
        "tags": ["Managers"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ManagerRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Manager created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Manager" } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Failed to create manager" }
        }
      }
    },
    "/api/managers/{id}": {
      "put": {
        "operationId": "updateManager",
        "summary": "Update a manager (should be authenticated)",
        "tags": ["Managers"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ManagerRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Manager updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Manager" } } }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Manager not found" },
          "500": { "description": "Failed to update manager" }
        }
      },
      "delete": {
        "operationId": "deleteManager",
        "summary": "Delete a manager (should be authenticated)",
        "tags": ["Managers"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Manager deleted",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Manager not found" },
          "500": { "description": "Failed to delete manager" }
        }
      }
    },
    "/api/events": {
      "get": {
        "operationId": "getEvents",
        "summary": "List events",
        "tags": ["Events"],
        "parameters": [
          { "name": "lang", "in": "query", "required": false, "schema": { "type": "string", "enum": ["ru", "kk", "en"] } }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } } } }
          },
          "500": { "description": "Failed to fetch events" }
        }
      },
      "post": {
        "operationId": "createEvent",
        "summary": "Create an event (should be authenticated)",
        "tags": ["Events"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventWrapper" } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Failed to create event" }
        }
      }
    },
    "/api/events/{id}": {
      "get": {
        "operationId": "getEventById",
        "summary": "Get a single event",
        "tags": ["Events"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Event",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Event" } } }
          },
          "404": { "description": "Event not found" },
          "500": { "description": "Failed to fetch event details" }
        }
      },
      "put": {
        "operationId": "updateEvent",
        "summary": "Update an event (should be authenticated)",
        "tags": ["Events"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Event updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EventWrapper" } } }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Event not found" },
          "500": { "description": "Failed to update event" }
        }
      },
      "delete": {
        "operationId": "deleteEvent",
        "summary": "Delete an event (should be authenticated)",
        "tags": ["Events"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Event deleted",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Event not found" },
          "500": { "description": "Failed to delete event" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/auth/login. NOTE: not currently enforced by any server middleware."
      }
    },
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "Manager": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "role": { "type": "string" },
          "image_url": { "type": "string", "nullable": true },
          "photo_url": { "type": "string", "nullable": true },
          "language": { "type": "string" },
          "display_order": { "type": "integer", "format": "int64" },
          "reception_day": { "type": "string", "nullable": true },
          "reception_time": { "type": "string", "nullable": true }
        }
      },
      "ManagerRequest": {
        "type": "object",
        "required": ["name", "role"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "role": { "type": "string", "maxLength": 255 },
          "image_url": { "type": "string", "nullable": true },
          "language": { "type": "string", "enum": ["ru", "kk", "en"] },
          "display_order": { "type": "integer", "format": "int64" }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "event_date": { "type": "string", "format": "date" },
          "time": { "type": "string", "nullable": true },
          "location": { "type": "string", "nullable": true },
          "language": { "type": "string" },
          "news_id": { "type": "integer", "format": "int64", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "EventRequest": {
        "type": "object",
        "required": ["title", "event_date"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "event_date": { "type": "string", "format": "date" },
          "time": { "type": "string", "maxLength": 50, "nullable": true },
          "location": { "type": "string", "maxLength": 255, "nullable": true },
          "language": { "type": "string", "enum": ["ru", "kk", "en"] },
          "news_id": { "type": "integer", "format": "int64", "nullable": true }
        }
      },
      "EventWrapper": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "event": { "$ref": "#/components/schemas/Event" }
        }
      },
...
    }
  }
}